Completed

+++ components/Home.jsx
@@ -11,34 +11,46 @@ const Home = () => {
 
   const uid = user?.id;
   const fetchNotes = async (id) => {
-    const response = await fetch(`/api/note/${id}`);
-    const data = await response.json();
+    if (isSignedIn && isLoaded) {
+      const response = await fetch(`/api/note/${id}`);
+      const data = await response.json();
 
-    setNotes(data);
+      setNotes(data);
+    }
   };
-
-  const deleteNote = async (id) =>{
-    const response = await fetch(`/api/note/${id}`,{method: 'DELETE'});
-    const data = await response.json();
-    setNotes(data);
-
+  function addNote(newNote) {
+    setNotes(prevNotes => {
+      return [...prevNotes, newNote];
+    });
   }
+  const deleteNote = async (id) => {
+    if(isSignedIn && isLoaded)
+    {
+
+      const response = await fetch(`/api/note/${id}`, { method: "DELETE" });
+      const data = await response.json();
+      setNotes(data);
+    }else{
+      setNotes(prevNotes => {
+        return prevNotes.filter((noteItem, index) => {
+          return index !== id;
+        });
+      });
+    }
+  };
   useEffect(() => {
     if (isLoaded || isSignedIn) {
       fetchNotes(uid);
     }
-  }, [isLoaded, isSignedIn,notes]);
+  }, [isLoaded, isSignedIn, notes]);
 
   // console.log(notes);
   return (
     <div className="">
       <Header />
-      <CreateArea fetchNote={fetchNotes} />
+      <CreateArea fetchNote={fetchNotes} onAdd={addNote} />
       {isLoaded && isSignedIn ? null : (
-        <Note
-          title={"Tip"}
-          content={"Create an account to save and edit notes"}
-        />
+        <Note title={"Tip"} content={"Create an account to save  notes"} />
       )}
 
       {notes.map((noteItem, index) => {
@@ -47,9 +59,9 @@ const Home = () => {
             <Note
               key={index}
               id={noteItem._id}
+              sId={index}
               title={noteItem.title}
               content={noteItem.note}
-
               onDelete={deleteNote}
             />
           </div>
